# Conflicts:
#	lib/services/db.dart
#	lib/widgets/group/group_shopping_list_detail.dart
#	lib/widgets/shoppinglist/create_shopping_list_view.dart
#	lib/widgets/shoppinglist/search_items_view_new.dart
#	lib/widgets/shoppinglist/shopping_list_detail.dart
#	lib/widgets/sidebar.dart
#	pubspec.yaml

+++ lib/models/Item.dart
@@ -34,7 +34,7 @@ class Item {
       category: data["category"],
       count: data["count"],
       bought: data["bought"],
-      prize: data["prize"],
+      prize: data["prize"].toDouble(),
     );
   }
 
@@ -46,7 +46,7 @@ class Item {
       category: data["category"],
       count: data["count"],
       bought: data["bought"],
-      prize: data["prize"],
+      prize: data["prize"].toDouble(),
     );
   }
